Fixed detaching windows. Changed a few names. Updated documentation.
Fixed the root window not receiving a pointer to the overlay it was created in.
Fixed windows retaining pointer to the overlay after being detached.
Moved `reindex` and `update_buffer` to be private methods of `Overlay` instead of nested functions.
Changed some names in the `WindowParams` struct.
Updated documentation of the public interface.